Boss stuffs

+++ NPCs/CaptainExplosiveBoss/CEDroneNPC.cs
@@ -6,17 +6,26 @@ using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 using static ExtraExplosives.GlobalMethods;
+using NPC = IL.Terraria.NPC;
 
 namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
 {
     public class CEDroneNPC : ModNPC
     {
         private Player _target = null;
-        private Vector2 endTarget;
+        private Vector2 _endTarget;
         private int _targetingFrames = 240 + Main.rand.Next(120) - 60;
         private bool _targetAquired = false;
+
+        private float[] _cordOffset = new float[]
+        {
+            Main.rand.NextFloat(-3f, 3f),
+            Main.rand.NextFloat(-15f, 15f)
+        };
+
+        private int spawnTimer = 60;
         // Speed vars, are more of a multiple than a strict value
-        private float speedX = 3f; // Used to control the x speed during phase 1
+        private float speedX = 7.5f; // Used to control the x speed during phase 1
 
         //Testing stuff
         private float[] _velocityStorage = new float[2];
@@ -43,15 +52,15 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
             npc.Hitbox = new Rectangle(0,0,32,32);
             npc.damage = 10;
             npc.defense = 5;
-            npc.lifeMax = 500;
+            npc.lifeMax = 150;
             npc.knockBackResist = 0f;
             npc.noTileCollide = true;
             npc.frame.Height = 22;
             npc.frame.Width = 22;
-            npc.noGravity = true;
+            npc.noGravity = false;
             npc.aiStyle = -1;
             npc.Center = new Vector2(11,11);
-
+            npc.rotation = Main.rand.Next(360);
             drawOffsetY = -8f;
         }
         
@@ -60,7 +69,7 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
         public override void FindFrame(int frameHeight)
         {
             npc.frameCounter++;
-            if (npc.frameCounter > 15)    //<--- updates per frame
+            if (npc.frameCounter > 10)    //<--- updates per frame
             {
                 npc.frameCounter = 0;
                 npc.frame.Y += frameHeight;
@@ -148,16 +157,37 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
 
         public override bool PreAI()
         {
-
-           // if (Keyboard.GetState().PressingShift() && testTimer <= 0)
+            npc.spriteDirection = npc.velocity.X > 0 ? 1 : -1;    // NPC will face the direction its moving
+            // if (Keyboard.GetState().PressingShift() && testTimer <= 0)
             //{
-                
-              //  Main.NewText($"Vectors {npc.position.X},{npc.position.Y} & {_target.position.X},{_target.position.Y} with single of 0 resulted in {ans}");    // debug info
+
+            //  Main.NewText($"Vectors {npc.position.X},{npc.position.Y} & {_target.position.X},{_target.position.Y} with single of 0 resulted in {ans}");    // debug info
             //}
-            
+            if (spawnTimer-- == 30)
+            {
+                //npc.velocity = Main.npc[npc.target].velocity;
+                return false;
+            }
+            else if (spawnTimer-- > 0)
+            {
+                npc.velocity.Y -= 0.05f;
+                //npc.velocity.X += (npc.ai[0] - 1) * 3;
+                return false;
+            }
+            else if (spawnTimer-- == 0)
+            {
+                Main.NewText("Gravity", Color.Cyan);
+                npc.noGravity = true;
+                npc.velocity.Y = 0;
+                npc.rotation = 0;
+            }
+
+            npc.velocity.X *= 0.37f;
+            npc.rotation = npc.velocity.X / 15f;    // Rotate the sprite based on x velocity (makes movement seem convincing)
             if (_target == null)
             {
                 npc.TargetClosest();
+               // Main.NewText($"The target is {npc.HasPlayerTarget}");
                 _target = Main.player[npc.target];
                 npc.velocity.Y -= 16 / Vector2.Distance(npc.position, _target.position) / 16f;
             }
@@ -168,37 +198,40 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
 
             if (_targetingFrames > 0)    // Targeting frames, the drone just hovers and locks in on the player
             {
+                float dist = npc.position.Y - (_target.position.Y + _cordOffset[1]);
                 _targetingFrames--;
-                if (npc.position.Y > _target.position.Y)        // catch in case the drone is below the player, somehow
+                if (npc.position.Y > (_target.position.Y + _cordOffset[1]) - 256)        // catch in case the drone is below the player, somehow
                 {
-                    npc.velocity.Y -= 0.3f;   
+                    npc.velocity.Y -= 0.5f;   
                 }
-                else if (Vector2.Distance(npc.position, _target.position) / 16f < 16 && _target.velocity.Y < 0)    // Raises / Lowers the drone so it is comfortably above the player
+
+                
+                //else if (dist < 32)    // Raises / Lowers the drone so it is comfortably above the player
+                //{
+                //    npc.velocity.Y -= Math.Abs(_target.velocity.Y)/8f;    // changes the y value so it stays near the player
+                //}
+                else if (dist > 60)
                 {
-                    npc.velocity.Y += _target.velocity.Y / 16f;    // changes the y value so it stays near the player
-                }
-                else if (Vector2.Distance(npc.position, _target.position) / 16f > 24)
-                {
-                    npc.velocity.Y += 0.2f; // changes the y value so it stays near the player
+                    npc.velocity.Y += 0.09f; // changes the y value so it stays near the player
                 }
                 else
                 {
                     npc.velocity.Y *= 0.75f;    // if its above the player by the set amount slowly lower the y velocity
                 }
-                Vector2 direction = (_target.Center - npc.Center).SafeNormalize(Vector2.UnitX);
-                if (npc.position.X > _target.position.X)
-                {
-                    npc.position.X += direction.X * speedX;
-                }
-                else
-                {
-                    npc.position.X += direction.X * speedX;
-                }
+                //Vector2 direction = (_target.Center - npc.Center).SafeNormalize(Vector2.UnitX);
+                //if (npc.position.X > _target.position.X)
+               // {
+                    npc.velocity.X += ((_target.Center - npc.Center).SafeNormalize(Vector2.UnitX).X + (_cordOffset[0]/10)) * speedX;
+                //}
+                //else
+               // {
+                   // npc.position.X -= direction.X * speedX;
+               // }
             }
 
             if (_targetingFrames <= 0)
             {
-                endTarget = _linearInterpolation(_target.position, npc.position);    // gets the cords of its target point
+                _endTarget = _linearInterpolation(_target.position, npc.position);    // gets the cords of its target point
                 _targetAquired = true;
                 _targetingFrames = 30;
             }
@@ -208,34 +241,68 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
         public override void AI()
         {
             if(_target.position.Y < npc.position.Y && Vector2.Distance(npc.position, _target.position)/16f < 8) Kill();
-            Vector2 direction = Vector2.Normalize(npc.position - endTarget);
-            if (_targetingFrames > 0)    // first step, will only really do anything if its high is up
+            Vector2 direction = Vector2.Normalize(npc.position - _endTarget);
+            if (_targetingFrames > 0)    
             {
                 _targetingFrames--;
-                npc.velocity.X += direction.X/20f;       
-                npc.velocity.Y += direction.Y/12f;
+                npc.velocity.X += direction.X * 8; 
+                if (_target.position.Y < npc.position.Y)
+                {
+                    npc.velocity.Y += direction.Y/20f;
+                }
+                else
+                {
+                    npc.velocity.Y += Vector2.Subtract(Vector2.Zero, direction).Y/20f;
+                }
             }
 
             if (_targetingFrames == 0)
             {
                 _targetingFrames = -1;
-                endTarget = _linearInterpolation(_target.position, npc.position);    // reaquire the target for better accuracy
+                _endTarget = _linearInterpolation(_target.position, npc.position);    // reaquire the target for better accuracy
             }
 
             if (_targetingFrames < 0)    // Accelerate the first step for a more 'energetic' attack
             {
-                npc.velocity.X += direction.X/3f;        
-                npc.velocity.Y += direction.Y/3f;
+                npc.velocity.X += direction.X * 8;        
+                if (_target.position.Y < npc.position.Y)
+                {
+                    npc.velocity.Y += direction.Y/3f;
+                }
+                else
+                {
+                    npc.velocity.Y -= Vector2.Subtract(Vector2.Zero, direction).Y/3f;
+                }
             }
         }
 
         public override void PostAI()
         {
             npc.FindFrame();
+            if (spawnTimer <= 0) return;
+            float rotation = npc.rotation % 360;
+            
+            
+            if (rotation != 0)
+            {
+                if (rotation > 90 && rotation < 270) // rotated between 1 and 180 degrees
+                {
+                   // Main.NewText($"Rotating: {(float)Math.Cos(npc.rotation/180 * 2 * Math.PI)/10}");
+                    npc.rotation -= (float)Math.Cos(npc.rotation/180 * 2 * Math.PI)/5;
+                }
+                else  // rotated between 180 and 359 degrees
+                {
+                   // Main.NewText($"Rotating: {(float)Math.Sin(npc.rotation/180 * 2 * Math.PI)/10}");
+                    npc.rotation += (float)Math.Sin(npc.rotation/180 * 2 * Math.PI)/5;
+                }
+            }
+            
+            //Main.NewText($"Rotation: {npc.rotation}", Color.Aquamarine);
         }
 
         public override void OnHitPlayer(Player target, int damage, bool crit)
         {
+            if (spawnTimer > 0) return;
             Kill();    // on collide kill it (here so it wont deal damage)
         }
 
@@ -324,7 +391,7 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
             base.NPCLoot();
         }
 
-        private void _moveEvent(int playerId)
+        /*private void _moveEvent(int playerId)
         {
             float dist = Vector2.Distance(Main.player[playerId].position, npc.position) / 16f;
             float distX = (Main.player[playerId].position.X - npc.position.X);
@@ -341,7 +408,7 @@ namespace ExtraExplosives.NPCs.CaptainExplosiveBoss
             {
                 Main.NewText($"Distance found to be: {dist} which is X{distX}, Y:{distY}", Color.Purple);
             }
-        }
+        }*/
         
         /// <summary>
         /// Uses linear interpolation to find the y intercept of a line which passes through 2 points a and b
